Added an experimental transport that uses the new hotplug features available in libusb >=1.0.16.
Warning - there is an issue in libusb < 1.0.18 that can cause segfaults within the event handler
when unplugging devices (if the event handler is delayed for some reason).

Implemented an Instrument class that provides a general purpose instrument driver based on the
device protocols. Eventually the Camera class will be able to derive from this, when it is safe
to merge the hotplugging transport in properly.