Flash When setting a connection string for libserialport it does not handle symlinks correctly (yet) so when trying to use a /dev/serial/by-id path it will fail to connect. This change uses std::filesystem to return the canonical path to use instead.
Bayer
The decoder was using std::async to run the Even and Odd decoding in separate threads. Firstly it only
needed to run one of those since the other could be run on the calling thread. Secondly std::async under
gcc/clang will not use a thread pool so new threads are created every time this is called. The emergent
library has been updated to include a PersistentThread (used by its own thread pool) which is now used
by Bayer so that a single thread is created and retained; this results in one extra thread per thread calling
the Bayer static function.